+++ NEWS.md
@@ -1,4 +1,5 @@
 ### httptest 1.2.1 (under development)
+* New context `capture_requests` to collect the responses from real requests and store them as mock files
 * `mockDownload` now processes request URLs as `mockRequest` does
 
 ## httptest 1.2.0